My new branch (#37)
* Minor changes...

Put in title and words for image...

* Update vuepress/docs/refarch/README.md

Just a full stop...

Co-authored-by: Miguel de Barros <miguel@debarros.me>

* Changes to ML 2.0 Ref. Arch. Document - 20210816a

The latest changes to my document include some layout updates and an image.

* Updates for ML 2.0 Ref. Arch. Document - 20210816b

* tweaks to reference architecture documentation section

- fixed lint issues
- added plugins for markdown parsing and footnotes
- fixed tables
- fixed footnotes

* Removed duplicate interfaces.md

* Final changes push - 08/24/2021

* Transfers BC: Draft page copy complete - 08/25/2021

Please advise if any amendments are required.

* Accounts Lookup & Discovery BC - 20210825a

* Minor change - 20210825

* Minor change - replaced image - 20210825

aldPartyParticipantDisassoc_20210825.png

* Minor change - wording - 20210825

Added "current" to 5th point in Concluding Comments.

* Minor change - Hid "Notes" - 20210825

Hid the Notes section as currently empty.

* Updates - Accounts Lookup & Discovery - 20210825

Added ALD page and made a number of changes.

* Updates - ALD & Quoting

Updates for the Account Lookup & Discovery BC, as well as the Quotes & Agreements BC.

* Minor change - 20210826

Moved "Notes"

* Minor change - 20210826

Updated order at the base of the file.

* Minor change - 20210826

Commented "Notes"

* BC Update - accountsAndBalances - 20210826

Completed draft BC layout and content.

* Settlements and Transfers Tweaks

Draft version of the Settlements and Transfers BCs

* Various Updates - 20210827

Updates to Account Lookups & Discovery, Quoting & Agreements, Settlements, and Transfers BCs.

* Third-Party API BC Updates... more to follow.

* New updates - 20210830a

* MyNewBranch Updates - 20210830b

* Added Intro directory (in honor of Pedro)

* 3rd-Party API BC Updates

I have completed a number of the UCs, working on the rest.  I will rewrite the 3rd-Party Initiated Transaction Request so that it is similar to the PISP Bulk Transaction Request as they are very similar.  This will most likely be completed Monday, after which I will start the Audit BC.

* 3rd-Party Assets...

Images for the 3rd-Party BC.

* Removing introduction.md... again...

* Various changes for 3rd-party API BC.

* Minor change - one more to follow.

* Final update for draft version of Third-Party API BC_09172021

The update contains all of the UCs for both the Linking and Transaction scenarios.

* Update docs/refarch/boundedContexts/thirdPartyApi/index.md

Co-authored-by: Miguel de Barros <miguel@debarros.me>

* Update docs/refarch/boundedContexts/thirdPartyApi/index.md

Good with this.  Thanks.

Co-authored-by: Miguel de Barros <miguel@debarros.me>

* Auditing BC: Text and images

Ready for reviewing and merging with the Main document.

* Update docs/refarch/boundedContexts/auditing/index.md

Co-authored-by: Miguel de Barros <miguel@debarros.me>

* Update docs/refarch/boundedContexts/auditing/index.md

Co-authored-by: Miguel de Barros <miguel@debarros.me>

* Update docs/refarch/boundedContexts/auditing/index.md

Co-authored-by: Miguel de Barros <miguel@debarros.me>

* Update docs/refarch/boundedContexts/auditing/index.md

Co-authored-by: Miguel de Barros <miguel@debarros.me>

* 3rd-Party API BC Update

I have updated a single Transfers UC in order to obtain feedback from the BC Owner regarding accuracy and suitability.  Updated transaction: Third-Party Initiated Transaction Request.

* Inserted Heading ID links (3rd-Party API BC)

* fixes to dependencies

* Third-Party API BC Tweak

I have made a small tweak to the Use Cases Scenarios description to get it to update the published version.  This may be a duplicate.

* Account Lookup & Discovery BC Flows V2

I have updated the overall and UC descriptions for the BC.

* PR Updates

Some updates to the various BCs that I am busy with...

* thirdPartyApi Tweaks

I have implemented a number of tweaks to the thirdPartyAPI BC UCs.  Please review and revert with any issues.  Thanks.

* Third Party Update...

* Third Party API BC

One final little tweak...

Co-authored-by: Miguel de Barros <miguel@debarros.me>