Accounts & Balances Update (#60)
* Accounts & Balances Update

I have completed the tweaks that were required for the Accounts & Balances BC documentation.  Due to the universality of the BC, certain of the UCs make use of UC workflows from other BCs, notably, Participant Lifecycle, Settlements, and Transfers.  There is still work to be done in those BCs, but there have been some updates to them through the requirements of this BC.  Please note this as you review and comment.  Thanks.

* Accounts & Balances - Minor Tweak

I esc'd the parentheses around the c in the Note: